feat(be):과복용 판단 기능 수정

+++ backend/yaksok/src/main/java/com/ssafy/yaksok/analyze/dto/IngredientOverdoseResult.java
@@ -5,8 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 /**
  * 성분별 과복용 판단 결과 DTO
  */
@@ -17,8 +15,8 @@ import java.math.BigDecimal;
 public class IngredientOverdoseResult {
     private Long ingredientId; // 성분 ID
     private String ingredientName; // 성분명
-    private BigDecimal totalAmount; // 총 섭취량
-    private BigDecimal maxIntakeValue; // 최대 섭취 기준
+    private Double totalAmount; // 총 섭취량
+    private Double maxIntakeValue; // 최대 섭취 기준
     private String unit; // 단위
     private boolean isOverdose; // 과복용 여부
 }
